backend/requirements.txt
# Core Framework
Flask==3.0.0
Flask-CORS==4.0.0
Flask-SQLAlchemy==3.1.1

# Database
SQLAlchemy==2.0.35

# Scheduling
APScheduler==3.10.4

# Calendar/ICS parsing
icalendar==5.0.11
python-dateutil==2.8.2

# HTTP requests
requests==2.31.0

# Environment variables
python-dotenv==1.0.0

# Email (for digest delivery)
email-validator==2.1.0

# Timezone handling
pytz==2023.3

# Data validation
marshmallow==3.20.1

# Development
pytest==7.4.3
pytest-flask==1.3.0
